requirements.txt
openai>=1.30.0
Pillow>=10.1.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
